+++ jschon/jsonpointer.py
@@ -70,6 +70,12 @@ class JSONPointer(Sequence[str]):
        and referred to as *keys* in the JSONPointer class.
     """
 
+    malformed_exc: Type[JSONPointerMalformedError] = JSONPointerMalformedError
+    """The exception to raise if the input is not a valid JSON Pointer"""
+
+    reference_exc: Type[JSONPointerReferenceError] = JSONPointerReferenceError
+    """The exception to raise if the JSON Pointer cannot be resolved against a document"""
+
     _json_pointer_re = re.compile(JSON_POINTER_RE)
     _array_index_re = re.compile(JSON_INDEX_RE)
 
@@ -79,7 +85,7 @@ class JSONPointer(Sequence[str]):
 
         :param values: each value may either be an RFC 6901 string, or an iterable
             of unescaped keys
-        :raise JSONPointerMalformedError: if a string argument does not conform to
+        :raise cls.malformed_exc: if a string argument does not conform to
             the RFC 6901 syntax
         """
         self = object.__new__(cls)
@@ -88,7 +94,7 @@ class JSONPointer(Sequence[str]):
         for value in values:
             if isinstance(value, str):
                 if not JSONPointer._json_pointer_re.fullmatch(value):
-                    raise JSONPointerMalformedError(f"'{value}' is not a valid JSON pointer")
+                    raise cls.malformed_exc(f"'{value}' is not a valid JSON pointer")
                 self._keys.extend(self.unescape(token) for token in value.split('/')[1:])
 
             elif isinstance(value, JSONPointer):
@@ -184,7 +190,7 @@ class JSONPointer(Sequence[str]):
         will always fail.
 
         :param document: any Python object
-        :raise JSONPointerReferenceError: if `self` references a non-existent
+        :raise self.reference_exc: if `self` references a non-existent
             location in `document`
         """
 
@@ -208,7 +214,7 @@ class JSONPointer(Sequence[str]):
             except (KeyError, IndexError):
                 pass
 
-            raise JSONPointerReferenceError(f"Path '{self}' not found in document")
+            raise self.reference_exc(f"Path '{self}' not found in document")
 
         return resolve(document, collections.deque(self._keys))
 
@@ -266,6 +272,15 @@ class JSONPointer(Sequence[str]):
 
 
 class RelativeJSONPointer:
+    malformed_exc: Type[
+        RelativeJSONPointerMalformedError
+    ] = RelativeJSONPointerMalformedError
+    """The exception to raise if the input is not a valid Relative JSON Pointer"""
+    reference_exc: Type[
+        RelativeJSONPointerReferenceError
+    ] = RelativeJSONPointerReferenceError
+    """The exception to raise if the Relative JSON Pointer cannot be resolved against a document"""
+
     _regex = re.compile(RELATIVE_JSON_POINTER_RE)
 
     def __new__(
@@ -288,13 +303,13 @@ class RelativeJSONPointer:
             a value of 0, which is not allowed by the grammar, is treated as if
             there is no adjustment.
         :param ref: a :class:`JSONPointer` instance, or the literal ``'#'``
-        :raise RelativeJSONPointerMalformedError: for any invalid arguments
+        :raise cls.malformed_exc: for any invalid arguments
         """
         self = object.__new__(cls)
 
         if value is not None:
             if not (match := RelativeJSONPointer._regex.fullmatch(value)):
-                raise RelativeJSONPointerMalformedError(f"'{value}' is not a valid relative JSON pointer")
+                raise cls.malformed_exc(f"'{value}' is not a valid relative JSON pointer")
 
             up, over, ref = match.group('up', 'over', 'ref')
             self.up = int(up)
@@ -339,28 +354,37 @@ class RelativeJSONPointer:
         return f'RelativeJSONPointer({str(self)!r})'
 
     def evaluate(self, document: JSON) -> Union[int, str, JSON]:
+        """Return the value within `document` at the location referenced by `self`.
+
+        `document` must be an instance of :class:`JSON`, as evaluation relies
+        on parent and sibling links provided by that class.
+
+        :param document: a :class:`JSON` instance representing the document
+        :raise self.reference_exc: if `self` references a non-existent
+            location in `document`
+        """
         node = document
         for _ in range(self.up):
             if node.parent is None:
-                raise RelativeJSONPointerReferenceError('Up too many levels')
+                raise self.reference_exc('Up too many levels')
             node = node.parent
 
         if self.over:
             if node.parent is None:
-                raise RelativeJSONPointerReferenceError('No containing node for index adjustment')
+                raise self.reference_exc('No containing node for index adjustment')
             if node.parent.type != "array":
-                raise RelativeJSONPointerReferenceError(f'Index adjustment not valid for type {node.parent.type}')
+                raise self.reference_exc(f'Index adjustment not valid for type {node.parent.type}')
             adjusted = int(node.key) + self.over
             if adjusted < 0 or adjusted >= len(node.parent):
-                raise RelativeJSONPointerReferenceError(f'Index adjustment out of range')
+                raise self.reference_exc(f'Index adjustment out of range')
             node = node.parent[adjusted]
 
         if self.index:
             if node.parent is None:
-                raise RelativeJSONPointerReferenceError('No containing node')
+                raise self.reference_exc('No containing node')
             return int(node.key) if node.parent.type == "array" else node.key
 
         try:
             return self.path.evaluate(node)
         except JSONPointerReferenceError as e:
-            raise RelativeJSONPointerReferenceError from e
+            raise self.reference_exc from e